Shard the legacy-migration test so its runs parallelise
`test/lib/db/legacy-migration.test.ts` held six tests in one file, each running
a full legacy→current `initDb()` upgrade, ~13s sequential. Same recipe as
`test/lib/db/migration-restore/`: the legacy schema, the fixtures, and the
shared assertions move into `legacy-migration/helpers.ts`, and the six tests
split across three thematic shard files (backfill, questions, partial-state).
Each shard owns its own file-DB tracking + cleanup via a small
`createLegacyMigrationHarness()`, so `deno test --parallel` runs the three
files at the same time instead of one long serial file. No assertions changed.
